// tina/types.ts
declare type Field<WithNamespace extends boolean = false> = (
  | StringField
  | NumberField
  | BooleanField
  | DateTimeField
  | ImageField
  | ReferenceField
  | RichTextField<WithNamespace>
  | ObjectField<WithNamespace>
) &
  MaybeNamespace<WithNamespace>;
export declare type TinaField<WithNamespace extends boolean = false> =
  Field<WithNamespace> & MaybeNamespace<WithNamespace>;
declare type MaybeNamespace<WithNamespace extends boolean = false> =
  WithNamespace extends true
    ? {
        namespace: string[];
      }
    : {};
// ...
export interface BaseCollection {
  label?: string;
  name: string;
  path: string;
  indexes?: IndexType[];
  format?: "json" | "md" | "markdown" | "mdx" | "yaml" | "yml" | "toml";
  ui?: UICollection;
  /**
   * @deprecated - use `ui.defaultItem` on the each `template` instead
   */
  defaultItem?: DefaultItem<Record<string, any>>;
  /**
   * This format will be used to parse the markdown frontmatter
   */
  frontmatterFormat?: "yaml" | "toml" | "json";
  /**
   * The delimiters used to parse the frontmatter.
   */
  frontmatterDelimiters?: [string, string] | string;
  match?: {
    include?: string;
    exclude?: string;
  };
}
// ...
export declare type FieldCollection<WithNamespace extends boolean = false> = {
  /**
   * Fields define the shape of the content and the user input.
   *
   * https://tina.io/docs/reference/fields/
   */
  fields: TinaField<WithNamespace>[];
  templates?: undefined;
} & BaseCollection &
  MaybeNamespace<WithNamespace>;
declare type Document = {
  _sys: {
    title?: string;
    template: string;
    breadcrumbs: string[];
    path: string;
    basename: string;
    relativePath: string;
    filename: string;
    extension: string;
  };
};
export interface UICollection {
  /**
   * Customize the way filenames are generated during content creation
   */
  filename?: {
    /**
     * A callback which receives form values as an argument. The return value
     * here will be used as the filename (the extension is not necessary)
     *
     * eg:
     * ```ts
     * slugify: (values) => values.title.toLowerCase().split(" ").join("-")
     * ```
     */
    slugify?: (values: Record<string, any>) => string;
    /**
     * When set to `true`, editors won't be able to modify the filename
     */
    readonly?: boolean;
  };
  /**
   * Determines whether or not this collection can accept new docments
   * or allow documents to be deleted from the CMS.
   */
  allowedActions?: {
    create?: boolean;
    delete?: boolean;
  };
  /**
   * Forms for this collection will be editable from the global sidebar rather than the form panel
   */
  global?:
    | boolean
    | {
        icon?: any;
        layout: "fullscreen" | "popup";
      };
  /**
   * Provide the path that your document is viewable on your site
   *
   * eg:
   * ```ts
   * router: ({ document }) => {
   *   return `blog-posts/${document._sys.filename}`;
   * }
   * ```
   */
  router?: (args: {
    document: Document;
    collection: Collection<true>;
  }) => string | undefined;
}
export declare type DefaultItem<ReturnType> = ReturnType | (() => ReturnType);
declare type IndexType = {
  name: string;
  fields: {
    name: string;
  }[];
};